Update dependencies and add mockinterceptor
