--- assets/json/user/login.json
@@ -1,3 +1,0 @@
-{
-    "access_token": "my-access-token"
-}
