+++ lib/presentation/providers/providers.dart
@@ -3,7 +3,7 @@ import 'package:flap_app/data/repository/analytics/analytics_platform_manager.da
 import 'package:flap_app/data/repository/auth/auth_repository_impl.dart';
 import 'package:flap_app/data/repository/recipe/recipe_repository_impl.dart';
 import 'package:flap_app/data/repository/secure_storage/secure_storage_impl.dart';
-import 'package:flap_app/data/source/network/mockinterceptor.dart';
+import 'package:flap_app/data/source/network/mock/mockinterceptor.dart';
 import 'package:flap_app/data/source/network/spoonacular_api.dart';
 import 'package:flap_app/data/source/network/spoonacular_api_impl.dart';
 import 'package:flap_app/domain/repository/analytics/analytics_platform.dart';
